tests/wpt/tests/soft-navigation-heuristics/multiple-paint-entries-buffered.tentative.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Detect multiple soft navigations and ensure they buffer paint entries.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/soft-navigation-helper.js"></script>
</head>
<body>
  <a id=link>Click me!</a>
  <main id=main>
  </main>
  <script>
    testSoftNavigation({
      testName: "Test that multiple soft navigation buffer entries",
      numClicks: 2,
      extraValidations: async (t) => {
        const sns = await getBufferedEntries('soft-navigation');
        assert_equals(sns.length, 2);
        const icps = await getBufferedEntries('interaction-contentful-paint');
        assert_equals(icps.length, 2);
      },
    });
  </script>
</body>
</html>
